ProjektSieciGlebokie/requirements.txt
# Biblioteki do pobierania i wczytywania danych
kagglehub>=0.3.0
numpy>=1.22.0
Pillow>=9.0.0
scikit-learn>=1.0.0

# Biblioteki do sieci neuronowych (wybierz jedną opcję)
# TensorFlow/Keras (zalecane dla początkujących)
tensorflow>=2.10.0

# Alternatywnie PyTorch (odkomentuj jeśli wolisz PyTorch)
# torch>=1.12.0
# torchvision>=0.13.0

# Biblioteki do wizualizacji
matplotlib>=3.5.0

# Opcjonalne - przydatne do analizy danych
pandas>=1.4.0

